+++ bot_go_brr/src/drive.rs
@@ -47,7 +47,7 @@ impl DriveArg {
         }
     }
 
-    pub fn log(&self, tick: &u128) {
+    pub fn log(&self, tick: &u32) {
         use crate::utils::Log::*;
         let (name, button, precise) = self.to_strings();
         Base(
